Agrego: 	Fantasma.java 	Game.java
Modfico:
	Pacman.java (deja de inicializar y pasa a inicializarlas "Game.java")
	Tablero.java(comentarios)
	Menu.java(comentarios)
