+++ Pacman/src/Pacman/Fantasma.java
@@ -1,0 +1,25 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Pacman;
+
+/**
+ *
+ * @author Alex
+ */
+public class Fantasma extends Thread{
+
+    public Fantasma() {
+    }
+
+    @Override
+    public void run() {
+        
+        
+    }
+    
+    
+    
+}
